pyproject.toml
[project]
name = "groqmind"
version = "0.1.0"
description = "AI assistant with streaming responses, conversation memory, and persona switching — powered by Groq"
readme = "README.md"
authors = [
    { name = "devie", email = "awaloeddin@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.134.0",
    "groq>=1.0.0",
    "python-dotenv>=1.2.1",
    "sse-starlette>=3.2.0",
    "uvicorn>=0.41.0",
]

[project.scripts]
groqmind = "groqmind.main:start"

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"
